+++ src/Symfony/Component/Workflow/Registry.php
@@ -66,6 +66,23 @@ class Registry
         return $matched;
     }
 
+    /**
+     * @param object $subject
+     *
+     * @return Workflow[]
+     */
+    public function all($subject): array
+    {
+        $matched = array();
+        foreach ($this->workflows as list($workflow, $supportStrategy)) {
+            if ($supportStrategy->supports($workflow, $subject)) {
+                $matched[] = $workflow;
+            }
+        }
+
+        return $matched;
+    }
+
     private function supports(WorkflowInterface $workflow, $supportStrategy, $subject, $workflowName): bool
     {
         if (null !== $workflowName && $workflowName !== $workflow->getName()) {