+++ src/Symfony/Component/Workflow/Tests/RegistryTest.php
@@ -81,6 +81,33 @@ class RegistryTest extends TestCase
         $this->assertSame('workflow1', $w1->getName());
     }
 
+    public function testAllWithOneMatchWithSuccess()
+    {
+        $workflows = $this->registry->all(new Subject1());
+        $this->assertInternalType('array', $workflows);
+        $this->assertCount(1, $workflows);
+        $this->assertInstanceOf(Workflow::class, $workflows[0]);
+        $this->assertSame('workflow1', $workflows[0]->getName());
+    }
+
+    public function testAllWithMultipleMatchWithSuccess()
+    {
+        $workflows = $this->registry->all(new Subject2());
+        $this->assertInternalType('array', $workflows);
+        $this->assertCount(2, $workflows);
+        $this->assertInstanceOf(Workflow::class, $workflows[0]);
+        $this->assertInstanceOf(Workflow::class, $workflows[1]);
+        $this->assertSame('workflow2', $workflows[0]->getName());
+        $this->assertSame('workflow3', $workflows[1]->getName());
+    }
+
+    public function testAllWithNoMatch()
+    {
+        $workflows = $this->registry->all(new \stdClass());
+        $this->assertInternalType('array', $workflows);
+        $this->assertCount(0, $workflows);
+    }
+
     /**
      * @group legacy
      */